requirements.txt
crawl4ai
pydantic
python-dotenv
beautifulsoup4
asyncio
urllib3
fastapi
uvicorn